Добавление карточки с информацией о текущей погоде.

+++ app/src/main/java/com/github/zharovvv/open/source/weather/app/network/dto/Weather.kt
@@ -1,3 +1,3 @@
 package com.github.zharovvv.open.source.weather.app.network.dto
 
-data class Weather(val id: Long, val main: String, val description: String, val icon: String)
+data class Weather(val id: Int, val main: String, val description: String, val icon: String)
